update: settings file changes
