+++ Resources.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c51b44b7bd4dd32439df3d43fc60aad6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
