+++ Resources/Settings/ReadyPlayerMeSettings.asset
@@ -9,8 +9,9 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3b50d26d65b1fce41a02e310760cd025, type: 3}
-  m_Name: AvatarLoaderSettings
+  m_Script: {fileID: 11500000, guid: 15d14b7ac4eee8c4492f5f0a34572c9f, type: 3}
+  m_Name: ReadyPlayerMeSettings
   m_EditorClassIdentifier: 
-  AvatarCachingEnabled: 1
-  AvatarConfig: {fileID: 0}
+  partnerSubdomain: liv
+  avatarLoaderSettings: {fileID: 11400000, guid: de84cec7367dec641a6ce55341b6ccbd,
+    type: 2}
